pyproject.toml
[project]
name = "uavcrew-mcp-server"
version = "2.0.0"
description = "MCP Gateway for UAVCrew AI agent access to customer data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "PyJWT[crypto]>=2.8.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mcp-gateway = "mcp_server.server:main"
mcp-gateway-dev = "mcp_server.server:dev"
uavcrew = "mcp_server.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]
